Use DESCRIPTION rather than lockfile for dependencies
Rather than pinning dependencies to a specific version
use a DESCRIPTION file to specify the dependenies
and have the installation determine up-to-date versions

This avoids issues with different versions of R not working
with specific versions of the packages.

A specific run will still generate a lock file which
can be used to recreate the exact run if required.